blackjack-crew/requirements.txt
# Project dependencies
crewai
crewai.tools>=0.1.0
python-dotenv
pydantic
langchain
openai
litellm
